Fix Bitbucket link handling for branches with slashes
Bitbucket has a bug where the "source view" link for a branch
which contains a slash results in a 404.

The link to the "branch view" works with names containing
a slash so we do this for branches with slashes in its names.

See https://jira.atlassian.com/browse/BCLOUD-14422 for more information.