Filter statuses in the github driver
GitHub keeps multiple versions of a check with the same label.
The API returns the list in reverse chronological order:
https://developer.github.com/v3/repos/statuses/#list-statuses-for-a-specific-ref.

The status resource in go-scm does not expose any information
that might help distinguish between versions of checks with
the same label.

When a client of go-scm receives a list of statuses that contain
duplicate labels, it has no way of finding out which one is the
most recent one.

Change the github driver to filter the statuses and only return
one (the most recent) for each label.

Fixes #79