refactor: Let GH driver return scm.ErrNotFound.
Support consistent error handling across providers when it comes to
resources that could not be found.
fake driver already behaves correspondingly. However since it does not
return a HTTP response it is not possible to have one condition for
a 'not found' error that works for both providers. This PR makes that
possible now.